Fixed issue with "NullPointerException" after building into .jar.
This was caused by the resource lookup process, which would only work if the files were organized into directories.
When packaged into a jar, the "new File(path).listFiles()" would throw the NullPointerException would occur.

This addition has the disadvantage of requiring the "org.springframework.spring-core" dependency, in order to be able to use the PathMatchingResourcePatternResolver class.